requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tableauserverclient==0.24.0
